% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{addAlpha}
\alias{addAlpha}
\title{add transparency to a color}
\usage{
addAlpha(col, alpha = 1)
}
\arguments{
\item{col}{the color, or vector of colors, you want to add transparency to}

\item{alpha}{the amount of transparency. ranges from 0 to 1 inclusive, where 0 is completely transparent and 1 is completely opaque}
}
\value{
color value with transparency added
}
\description{
from http://www.magesblog.com/2013/04/how-to-change-alpha-value-of-colours-in.html
}
\examples{
"red" \%>\% addAlpha(0.5)
viridis(5) \%>\% addAlpha(0.7)
compare:
x <- list(rnorm(50,5),rnorm(50,5),rnorm(50,5),rnorm(50,5),rnorm(50,5),rnorm(50,5),rnorm(50,5),rnorm(50,5),rnorm(50,5))
simple(x, point_col = plasma(10))
simple(x, point_col = plasma(10) \%>\% addAlpha(0.5))

}

