% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{bar}
\alias{bar}
\title{plot categorical x quantitative data as bar plot}
\usage{
bar(sim, lab = rep(c(), length(sim)), CI = F, SE = F,
  bar_color = "grey80", jitter = T, point_col = "#00000080",
  y_limits = NA, median = FALSE, point_size = 1, sample_size = TRUE,
  ...)
}
\arguments{
\item{sim}{list of vectors you want to plot}

\item{lab}{labels for each group}

\item{CI}{logical. should a 95% confidence interval be drawn for each group? defaults to F}

\item{SE}{logical. should the dtandard errors for each group be drawn? deafults to F}

\item{bar_color}{color of the bars. can be a vector where length(bar_color) = number of groups}

\item{jitter}{logical. show the data jittered? defaults to TRUE}

\item{point_col}{color of the jittered data points}

\item{median}{logical. draw median for each group}

\item{point_size}{size of data points. defaults to 1.0. increase to make the data points bigger.}

\item{sample_size}{logical. plot sample size under each bar?}

\item{...}{other arguments to be passed to par()}
}
\value{
matrix showing x axis positions of the bars
}
\description{
plots each vector in your list as a group as a bar plot with the data on top, jittered
}
\examples{
bar(list(rnorm(20,5),rnorm(15,5)),lab=c("A","B"),xlab="group",ylab="response",jitter=F)
bar(list(rnorm(20,5),rnorm(15,5)),CI=T,lab=c("A","B"),xlab="group",ylab="response",jitter=F,bar_color=viridis(3)[1:2])
bar(list(rnorm(20,5),rnorm(15,5),rnorm(200,50),rnorm(50,1),rnorm(34,19)),CI=T,bar_color=viridis(7)[1:5])
bar(list(rnorm(20,5),rnorm(15,5)),CI=T,lab=c("A","B"),xlab="group",ylab="response",jitter=T,bar_color=viridis(10)[c(3,7)],point_size=1.5)

}

