% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{beeStrip}
\alias{beeStrip}
\title{beeStrip}
\usage{
beeStrip(data, lab = rep(c(), length(data)), point_size = 1.4,
  beeMethod = "center", line_color = "red", line_width = 3, jitter = T,
  point_col = viridis(length(data) + 2)[1:length(data)],
  y_limits = c(min(unlist(data), na.rm = T), max(unlist(data), na.rm = T)),
  median = FALSE, rug = TRUE, sample_size = T, IQR = F, side = -1,
  ...)
}
\arguments{
\item{data}{list of groups you want to plot where length(data) = number of groups}

\item{lab}{labels for each group}

\item{point_size}{size of the individual data points. default is 1.4}

\item{beeMethod}{see beeswarm(). default is "center"}

\item{line_color}{color of line drawn at the mean for each group. default is "red"}

\item{line_width}{width of the line drawn at the mean. default is 3.0}

\item{jitter}{logical. should the data be jittered? defaults to TRUE}

\item{point_col}{vector of colors. defaults to viridis colors}

\item{y_limits}{you know this one}

\item{median}{logical. should a diamond for the median be plotted?}

\item{rug}{logical. should a rug plot be plotted on the right side of the plot?}

\item{sample_size}{logical. should the sample size be plotted beneath each group?}

\item{IQR}{logical. Should a line representing the IQR for each group be plotted?}

\item{...}{other arguments to par()}
}
\value{
None
}
\description{
plots each vector in your list as a group with the data forming a histogram; draws a line at the mean for eaach group
}
\examples{
beeStrip(list(rnorm(50,50,5),rnorm(30,40,6),rnorm(10,60,2),rnorm(60,50,10),rnorm(30,39,4)),point_col=wes_palette(5, name = "Zissou"),line_color="black",median=T)
beeStrip(list(iris \%>\% filter(Species=="setosa") \%>\% .$Sepal.Length, iris \%>\% filter(Species=="versicolor") \%>\% .$Sepal.Length, iris \%>\% filter(Species=="virginica") \%>\% .$Sepal.Length),lab=c("setosa","versicolor","virginica"),ylab="sepal length",main="beeStrip()",xlab="species")

}

