% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{beeStripBox}
\alias{beeStripBox}
\title{plot quantitative x categorical data}
\usage{
beeStripBox(data, group, lab = rep(c(), length(data)), point_size = 1.4,
  beeMethod = "center", line_width = 3, point_col = ifelse(is.list(data)
  \%>\% rep(20), viridis(length(data) + 1)[1:length(data)],
  viridis(nlevels(group) + 1)[1:nlevels(group)]),
  y_limits = c(ifelse(is.list(data), min(unlist(data)), min(data)),
  ifelse(is.list(data), max(unlist(data), max(unlist(data))))), mean = FALSE,
  sample_size = T, side = -1, stats = T, box_thickness = 0.2,
  box_color = FALSE, ...)
}
\arguments{
\item{data}{list of groups you want to plot, or, if group is not left blank, a data frame}

\item{group}{if you gave a dataframe for data, group is the name of the column containing group identities}

\item{lab}{labels for groups}

\item{point_size}{size of plotting characters. defaults to 1.4}

\item{beeMethod}{see beeswarm. defaults to "center"}

\item{line_width}{width of the line}

\item{point_col}{color of the each group. defaults to viridis colors}

\item{y_limits}{limits of the y axis}

\item{mean}{does nothing right now}

\item{sample_size}{logical. plot sample size under each group?}

\item{side}{defaults to -1. determines whether the histogram are stacked to the right or the left}

\item{stats}{logical. do an ANOVA and print the p-value to the plot?}

\item{box_thickness}{thickness of the lines that form the boxplot. defaults to 0.2}

\item{box_color}{defaults to FALSE. color of the lines that form the boxes}

\item{...}{other arguments to pass to par()}
}
\value{
ANOVA results
}
\description{
plots a histogram of your individual data points alongside a boxplot for each group
}
\examples{
data(iris); beeStripBox(iris$Sepal.Length,iris$Species,xlab="species",ylab="sepal length",main="beeStripBox() example")

}

