% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{beeStripMod}
\alias{beeStripMod}
\title{plot qualitative x quantitative as Tufte's boxplot + histogram of data in each group}
\usage{
beeStripMod(data, group, lab = rep(c(), length(data)), point_size = 1.4,
  beeMethod = "center", line_width = 3, point_col = ifelse(is.list(data)
  \%>\% rep(20), viridis(length(data) + 1)[1:length(data)],
  viridis(nlevels(group) + 1)[1:nlevels(group)]),
  y_limits = c(ifelse(is.list(data), min(unlist(data)), min(data)),
  ifelse(is.list(data), max(unlist(data), max(data)))), sample_size = T,
  side = -1, red_median = T, stats = T, ...)
}
\arguments{
\item{data}{either a column in a dataframe containing the raw data you wish to plot, or a list, where len(your_list) == number of groups you're plotting}

\item{group}{if `data` is a column in a dataframe, `group` is a column in the same dataframe giving the group that each observation in `data` belongs to. If `data` is a list, don't worry about this argument.}

\item{lab}{labels for the groups}

\item{line_width}{width of the lines for the modified boxplots}

\item{point_col}{color of the data points for each group. defaults to viridis colors}

\item{y_limits}{limits of the y axis}

\item{sample_size}{logical. If TRUE, print the sample size under each group}

\item{red_median}{logical. Should the median be red? If FALSE, defaults to the color of the group.}

\item{stats}{logical. print overall p value from anova?}

\item{...}{other arguments to par()}
}
\value{
None
}
\description{
plots a modified boxplot from Tufte for each group alongside the data as arranged as a histogram
}
\examples{
data(iris)
beeStripMod(iris$Sepal.Length,iris$Species,xlab="species",ylab="sepal length",main="beeStripMod() example")


}

