% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{cats_meow}
\alias{cats_meow}
\title{plot quantitative x categorical data various ways}
\usage{
cats_meow(y, x = NA, lab = NA, SEM = FALSE, CI = TRUE,
  box_thickness = 0.2, plot_data = T, colors = viridis(5)[1:4], ...)
}
\arguments{
\item{y}{either a list of the data you wish to plot, where length(data) == # of groups, or a column of a dataframe containing y axis values}

\item{x}{NA if y is a list. Otherwise a column in a dataframe containing the group labels for each observation in y}

\item{lab}{labels for groups}

\item{SEM}{draw s.e.m.'s as fences? defaults to FALSE}

\item{CI}{draw 95 percent CIs as fences? defaults to TRUE}

\item{box_thickness}{thickness of the boxes of your boxplots}

\item{plot_data}{logical. plot the data jittered alongside the boxplot? Defaults to TRUE}

\item{colors}{colors of the data points. defaults to viridis colors}

\item{...}{other arguments to pass to par()}
}
\value{
none
}
\description{
boxplot + mean with fences for CIs or s.e.m.s + data jittered
}
\examples{
data(iris); cats_meow(iris$Sepal.Length,iris$Species, ylab="sepal length", xlab = "species")

}

