% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{paired}
\alias{paired}
\title{plot paired data}
\usage{
paired(x, y, lab = NA, box_thickness = 0.2, plot_points = T,
  colors = c("#440154FF", "#21908CFF") \%>\% addAlpha(0.1),
  line_color = "grey20", ...)
}
\arguments{
\item{x}{NA if y is a list. Otherwise a column in a dataframe containing the group labels for each observation in y}

\item{y}{either a list of the data you wish to plot, where length(data) == # of groups, or a column of a dataframe containing y axis values}

\item{lab}{labels for groups}

\item{box_thickness}{thickness of the boxes of your boxplots}

\item{colors}{colors of the data points. defaults to viridis colors}

\item{...}{other arguments to pass to par()}

\item{SEM}{draw s.e.m.'s as fences? defaults to FALSE}

\item{CI}{draw 95 percent CIs as fences? defaults to TRUE}

\item{plot_data}{logical. plot the data jittered alongside the boxplot? Defaults to TRUE}
}
\value{
none
}
\description{
boxplot + paired data connected with points
}
\examples{
paired(rnorm(10,5), rnorm(10,7), lab = c("X", "X_paired"), ylab = "measurement", main="paired() example")

}

