% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{scatter}
\alias{scatter}
\title{plot quantitative x quantitative data as scatter plot}
\usage{
scatter(x, y, xlab = "", ylab = "", line = T, stats = TRUE,
  color = "black", line_col = "red", confidenceInterval = T,
  plottingCharacter = 16, rug = T, ...)
}
\arguments{
\item{x}{vector of data for the x axis}

\item{y}{vector of data for the y axis}

\item{xlab}{x label}

\item{ylab}{y label}

\item{line}{logical. draw regression line? defaults to T}

\item{stats}{logical. compute sample size, p-value for regression slope, and r squared value and print to the corner of the graph? defaults to TRUE}

\item{color}{color of data points. defaults to "black"}

\item{line_col}{color of regression line. defaults to "red"}

\item{confidenceInterval}{logical. plot confidence bands about the slope?}

\item{plottingCharacter}{type of plotting character. defaults to 16, i.e., solid point}

\item{rug}{logical. plot a rug plot along the x and y axes? defaults to TRUE}

\item{...}{other arguments to par()}
}
\value{
None
}
\description{
scatter plot with nice defaults. automatically tests for a linear association between your two variables, draws regression lines, prints stats to the plot, etc.
}
\examples{
scatter(rnorm(50,50,10),rnorm(50,30,3))
scatter(trees[,1],trees[,2],xlab="tree girth (in.)",ylab="tree height (ft.)",main="scatter() example")


}

