% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{paired}
\alias{paired}
\title{plot paired data}
\usage{
paired(x, y, lab = NA, box_thickness = 0.2, plot_points = T,
  colors = c("#440154FF", "#21908CFF"), line_color = "#00000050", ...)
}
\arguments{
\item{x}{first vector or data frame with at least two columns}

\item{y}{second vector. length(x) must equal length(y) for the data to be paired. If y is left blank, x must be a dataframe with at least two columns. paired() assumes that the first two columns are the paired data. If x and y are both dataframes, paired() assumes the paired data are the first columns of each dataframe.}

\item{lab}{labels for groups}

\item{box_thickness}{thickness of the boxes of your boxplots}

\item{colors}{colors of the data points. defaults to viridis colors}

\item{line_color}{color of the lines to connect paired observations}

\item{...}{other arguments to pass to par()}

\item{plot_data}{logical. plot the data be plotted as points at the tips of each line}
}
\value{
none
}
\description{
boxplot + paired data connected with points
}
\examples{
paired(rnorm(10,5), rnorm(10,7), lab = c("X", "X_paired"), ylab = "measurement", main="paired() example")

}

