% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{simple}
\alias{simple}
\title{line for mean + data}
\usage{
simple(data, grouping = NULL, lab = NA, point_size = 1.2,
  line_color = "red", line_width = 3, jitter = T,
  point_col = viridis(length(data) + 2)[1:length(data)],
  y_limits = c(min(unlist(data), na.rm = T), max(unlist(data), na.rm = T)),
  median = FALSE, rug = TRUE, sample_size = T, IQR = F, ...)
}
\arguments{
\item{data}{input your data as a list where length(list) = number of groups}

\item{lab}{labels for groups}

\item{point_size}{size of points}

\item{line_color}{color of line for mean}

\item{jitter}{defaults to T}

\item{point_col}{defeaults to viridis colors}

\item{median}{plot median? defaults to false}

\item{rug}{plot 1-D rug plot?}

\item{sample_size}{defaults to true}

\item{IQR}{include line for IQR? defaults to false}

\item{ylim}{y limits}
}
\description{
plots each vector in your list as a group with the data jittered; draws a line for the mean
}
\examples{
x <- list(rnorm(40,40,5),rnorm(20,35,2),rnorm(25,41,2))
}

